Split some functionality out of the main `rten_convert.converter` module
Split graph and node types, error types and attribute parsing out of the large
`converter.py` module into some smaller submodules.
